分数计算完成

+++ cloudok-business-service/src/main/java/com/cloudok/uc/po/MemberPO.java
@@ -45,7 +45,7 @@ public class MemberPO extends PO {
 	
 	private Long state;
 	
-	private Integer wi;
+	private Double wi;
 	
-	private Integer ti;
+	private Double ti;
 }
